+++ tests/api.test.js
@@ -7,9 +7,8 @@ const db = require('../src/config/db.config')
 const express = require('express')
 
 const rideRoutes = require('../src/routes/rides')
-const healthRoute = require('../src/routes/health')
 const app = express()
-app.use('/health', healthRoute)
+const resetDB = require('./setup')
 app.use('/rides', rideRoutes)
 
 const createRide = () => {
@@ -29,104 +28,16 @@ const rider = {
 describe('API tests', () => {
   beforeEach((done) => {
     // Resets the table for each test, also resets the auto_increment to zero
-    db.getDB().run(`DROP TABLE IF EXISTS rides`, function (err) {
-      if (err) {
-        return done(err)
-      }
-    })
-    db.getDB().serialize((err) => {
-      if (err) {
-        return done(err)
-      }
-      db.initSchema()
-    })
+    resetDB(db, done)
     done()
   })
 
-  describe('GET /health', () => {
-    it('should return health', (done) => {
-      request(app)
-        .get('/health')
-        .expect('Content-Type', /text/)
-        .expect(200, done)
-    })
-  })
-
   describe('Rides API', () => {
-    it('should return error message if there is no rides', async () => {
-      const response = await request(app).get('/rides').send()
-      assert.equal(response.body.error_code, 'RIDES_NOT_FOUND_ERROR')
-    })
     it('should return error message if the ride is not found', async () => {
       const response = await request(app).get('/rides/10').send()
       assert.equal(response.body.error_code, 'RIDES_NOT_FOUND_ERROR')
     })
-    it('can create a new ride with valid inputs', async () => {
-      const response = await request(app).post('/rides').send(rider)
-      assert.equal(response.body[0].rideID, 1)
-    })
-    it('throws an error if startLatitude and startLongitude is not valid', async () => {
-      // make sure the copy is deep copy
-      const invalidRider = JSON.parse(JSON.stringify(rider))
-      invalidRider.start_lat = 92
-      invalidRider.start_long = 200
-      const response = await request(app).post('/rides').send(invalidRider)
-      assert.equal(response.body.error_code, 'VALIDATION_ERROR')
-      assert.equal(
-        response.body.message,
-        'Start latitude and longitude must be between -90 - 90 and -180 to 180 degrees respectively'
-      )
-    })
 
-    it('throws an error if endLatitude and endLongitude is not valid', async () => {
-      const invalidRider = JSON.parse(JSON.stringify(rider))
-      invalidRider.end_lat = 99
-      invalidRider.end_long = 200
-      const response = await request(app).post('/rides').send(invalidRider)
-      assert.equal(response.body.error_code, 'VALIDATION_ERROR')
-      assert.equal(
-        response.body.message,
-        'End latitude and longitude must be between -90 - 90 and -180 to 180 degrees respectively'
-      )
-    })
-    it('throws an error if riderName is not valid', async () => {
-      const invalidRider = JSON.parse(JSON.stringify(rider))
-      invalidRider.rider_name = ''
-      const response = await request(app).post('/rides').send(invalidRider)
-      assert.equal(response.body.error_code, 'VALIDATION_ERROR')
-      assert.equal(
-        response.body.message,
-        'Rider name must be a non empty string'
-      )
-    })
-    it('throws an error if driverName is not valid', async () => {
-      const invalidRider = JSON.parse(JSON.stringify(rider))
-      invalidRider.driver_name = ''
-      const response = await request(app).post('/rides').send(invalidRider)
-      assert.equal(response.body.error_code, 'VALIDATION_ERROR')
-      assert.equal(
-        response.body.message,
-        'Driver name must be a non empty string'
-      )
-    })
-    it('throws an error if driverVehicle is not valid', async () => {
-      const invalidRider = JSON.parse(JSON.stringify(rider))
-      invalidRider.driver_vehicle = ''
-      const response = await request(app).post('/rides').send(invalidRider)
-      assert.equal(response.body.error_code, 'VALIDATION_ERROR')
-      assert.equal(
-        response.body.message,
-        'Driver vehicle must be a non empty string'
-      )
-    })
-
-    it('can fetch rides', async () => {
-      await createRide()
-      await createRide()
-      const response = await request(app).get('/rides').send()
-      assert.equal(response.body[0].rideID, 1)
-      assert.equal(response.body.length, 2)
-    })
     it('can get a specific ride', async () => {
       await createRide()
       await createRide()
@@ -139,20 +50,6 @@ describe('API tests', () => {
       const response = await request(app).get('/rides/2').send()
       assert.equal(response.body.error_code, 'RIDES_NOT_FOUND_ERROR')
       assert.equal(response.body.message, 'Could not find any rides')
-    })
-    it('can fetch rides on a specific page', async () => {
-      await createRide()
-      await createRide()
-      await createRide()
-      await createRide()
-      await createRide()
-      const response = await request(app).get('/rides?page=2').send()
-      assert.equal(response.body[0].rideID, 3)
-      assert.equal(response.body.length, 2)
-    })
-    it('should return error message if the page is empty', async () => {
-      const response = await request(app).get('/rides?page=2').send()
-      assert.equal(response.body.error_code, 'RIDES_NOT_FOUND_ERROR')
     })
   })
 })
